requirements.txt
fastapi[standard]
pydantic-settings
requests
pandas
uvicorn
pytest
pytest-dotenv
httpx
httpretty
